фfix(imperfections) outsource router to make App.tsx cleaner
